+++ assets/js/utils/functions.js
@@ -1,11 +1,12 @@
 import $ from 'jquery'
 
 export function configTimeago() {
+  $.timeago.settings.allowFuture = true;
   $.timeago.settings.strings = {
     prefixAgo: null,
     prefixFromNow: null,
     suffixAgo: 'ago',
-    suffixFromNow: 'from now',
+    suffixFromNow: 'remaining',
     seconds: 'less than a minute',
     minute: 'a minute',
     minutes: '%d minutes',