+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  * removed default registration mechanism
  * by default: switched from client and server to serverside-only form validation. De-validation on input change now global.
  * exeriment minTime now retrieved from config instead of hard-coded
+ * improved performance of timer and switched to timeago display to avoid displaying seconds on higher minutes
 
 ### 2.3.1
  * fixed port number in email links